refactor: forest-tool shed summarize-tipsets
